// apps/web/lib/validations/database.schemas.ts
export const jsonSchema: z.ZodSchema<Json> = z.lazy(() =>
  z
    .union([
      z.string(),
      z.number(),
      z.boolean(),
      z.record(z.string(), z.union([jsonSchema, z.undefined()])),
      z.array(jsonSchema),
    ])
    .nullable(),
);
// ...
export const publicConfigurationsInsertSchema = z.object({
  aws_account_id: z.string().optional().nullable(),
  aws_region: z.string().optional().nullable(),
  cloud_identity_id: z.string().optional().nullable(),
  cluster_id: z.string().optional().nullable(),
  container_platform: z.string(),
  create_vpc: z.boolean().optional().nullable(),
  created_at: z.string().optional().nullable(),
  db_max_capacity: z.number().optional().nullable(),
  db_min_capacity: z.number().optional().nullable(),
  description: z.string().optional().nullable(),
  dns_domain_name: z.string().optional().nullable(),
  dns_hosted_zone: z.string().optional().nullable(),
  download_count: z.number().optional().nullable(),
  eks_cluster_admins: z.string().optional().nullable(),
  enable_cloudfront_waf: z.boolean().optional().nullable(),
  enable_dns: z.boolean().optional().nullable(),
  enable_gitops_destination: z.boolean().optional().nullable(),
  enable_karpenter: z.boolean().optional().nullable(),
  enable_redis: z.boolean().optional().nullable(),
  environment_repository: z.string().optional().nullable(),
  environment_stage: z.string(),
  full_config: jsonSchema.optional().nullable(),
  gitops_app_template: z.string().optional().nullable(),
  gitops_app_token: z.string().optional().nullable(),
  gitops_argocd_token: z.string().optional().nullable(),
  gitops_destinations_repo: z.string().optional().nullable(),
  gitops_repository: z.string().optional().nullable(),
  id: z.string().optional(),
  last_downloaded_at: z.string().optional().nullable(),
  project_name: z.string(),
  redis_allowed_cidr_blocks: z.string().optional().nullable(),
  ses_queues_topics: z.string().optional().nullable(),
  status: z.string().optional().nullable(),
  terraform_version: z.string(),
  updated_at: z.string().optional().nullable(),
  user_id: z.string(),
  vpc_cidr: z.string().optional().nullable(),
});